requirements.txt
# The app itself

-e .


# Runtime dependencies

Django==1.8.9
django-cms==3.2.0


# Example app
django-bootstrap3==6.2.2


# CMS Plugins
djangocms-file==0.1
djangocms-inherit==0.1
djangocms-link==1.7.1
djangocms-picture==0.1
djangocms-snippet==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.8.0


# CMS Optional, recommended packages
Pillow==3.0.0
django-filer==1.0.4
cmsplugin-filer==1.0.0
django-reversion>=1.8,<1.9